feat: 用户组绑定所有用户
