+++ src/views/users/Group/UserGroupDetail/index.vue
@@ -9,11 +9,13 @@
 <script>
 import { GenericDetailPage } from '@/layout/components'
 import GroupInfo from './GroupInfo'
+import GroupUser from './GroupUser'
 
 export default {
   components: {
     GenericDetailPage,
-    GroupInfo
+    GroupInfo,
+    GroupUser
   },
   data() {
     return {
@@ -28,9 +30,7 @@ export default {
           },
           {
             title: this.$t('route.UserList'),
-            name: 'GroupAddUser'
-            // TODO 稍后加权限
-            // hidden: () => !this.$hasPerm('accounts.change_accounttemplate')
+            name: 'GroupUser'
           }
         ]
       }
